Move files into their correct directories according to architecture.
