+++ domain/build.gradle
@@ -54,6 +54,13 @@ dependencies {
 
     // Paging 3.0
     implementation "androidx.paging:paging-compose:$paging_version"
+
+    //Camerax
+    def camerax_version = "1.3.0-alpha04"
+    implementation "androidx.camera:camera-core:${camerax_version}"
+
+    //Barcode Analyser
+    implementation 'com.google.mlkit:barcode-scanning:17.0.3'
 }
 
 task jacocoTestReport(type: JacocoReport, dependsOn: ['testDebugUnitTest', 'createDebugCoverageReport']) {
